Change column from integer to biginteger

+++ app/models/stock.py
@@ -8,7 +8,7 @@ class Stock(db.Model):
     code = db.Column(db.String, nullable=False)
     name = db.Column(db.String, nullable=False)
     listing_date = db.Column(db.Date, nullable=False)
-    outstanding_shares = db.Column(db.Integer, nullable=False)
+    outstanding_shares = db.Column(db.BigInteger, nullable=False)
     prices = db.relationship('StockPrice', backref='stock', lazy=True)
     created_at = db.Column(db.DateTime, default=db.func.current_timestamp())
     updated_at = db.Column(db.DateTime, default=db.func.current_timestamp(), onupdate=db.func.current_timestamp())
